pyproject.toml
[tool.poetry]
name = "ai-deepfake-detector"
version = "0.1.0"
description = ""
authors = ["ohsmaan"]
readme = ""
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
# Add your main dependencies here, for example:
torch = "^2.7.1"
torchvision = "^0.22.1"
torchaudio = "2.7.1"
opencv-python = "4.11.0.86"
scikit-learn = "1.7.0"
sentence-transformers = "4.1.0"
transformers = "4.52.4"
uvicorn = "^0.34.3"
python-multipart = "^0.0.20"
anthropic = "^0.54.0"
# Additional dependencies for video processing
aiofiles = "^23.2.1"
numpy = "^2.3.1"
pillow = "^11.2.1"
fastapi = "^0.115.13"
huggingface-hub = "^0.33.0"
python-dotenv = "^1.1.0"
requests = "^2.32.4"
soundfile = "^0.12.1"
scipy = "^1.14.0"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.2.0"
mypy = "^1.2.0"
flake8 = "^6.0.0"

